+++ test.cpp
@@ -1,0 +1,7 @@
+#include"store.h"
+#include"node.h"
+
+int main(){
+    
+    return 0;
+}
